Revert "protocol fixes"
This reverts commit dd0cdf587ef92ea9decf6fae637daf620df0c4d1.

+++ server/session/handler_enchanting.go
@@ -111,7 +111,7 @@ func (s *Session) sendEnchantmentOptions(tx *world.Tx, c Controllable, pos cube.
 		for _, enchant := range selectedEnchants[i] {
 			id, _ := item.EnchantmentID(enchant.Type())
 			enchants = append(enchants, protocol.EnchantmentInstance{
-				Type:  uint32(id),
+				Type:  byte(id),
 				Level: byte(enchant.Level()),
 			})
 		}
